requirements.txt
requests
pandas
numpy
scikit-learn
Flask
Flask-CORS
gradio
python-dotenv
pyngrok
